timiiiing

+++ src/instr_impl.h
@@ -4,7 +4,7 @@
 #include "cpu.h"
 #include "instr_info.h"
 
-typedef int (*instr_impl_fn)(struct cpu *cpu, const struct instr_info *info);
+typedef uint16_t (*instr_impl_fn)(struct cpu *cpu, const struct instr_info *info);
 
 extern const instr_impl_fn INSTR_IMPL[256];
 extern const instr_impl_fn INSTR_IMPL_PREFIX[256];
